coffee-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
pydantic==2.9.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai==1.50.0
httpx==0.27.0
python-multipart==0.0.9
playwright==1.49.1
cryptography==44.0.0
structlog==24.4.0
google-auth>=2.20.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
anthropic>=0.39.0
reportlab>=4.0.0
aiosmtplib>=3.0.0
